pyproject.toml
[tool]
[tool.poetry]
name = "centraal_dataframework"
version = "0.1"
homepage = "https://github.com/centraal-api/centraal-dataframework"
description = "Utilidades para interactuar con Azure Datalake."
authors = ["centraal.studio <equipo@centraal.studio>"]
readme = "README.md"
license =  "Apache-2.0"
classifiers=[
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: Apache Software License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
]
packages = [
    { include = "centraal_dataframework" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
black  = { version = "^22.3.0", optional = true}
isort  = { version = "^5.8.0", optional = true}
flake8  = { version = ">=3.9.2", optional = true}
flake8-docstrings = { version = "^1.6.0", optional = true }
pytest  = { version = ">=6.2.4", optional = true}
pytest-cov  = { version = ">=2.12.0", optional = true}
tox  = { version = "^3.20.1", optional = true}
virtualenv  = { version = "^20.2.2", optional = true}
pip  = { version = ">=22.0", optional = true}
mkdocs  = { version = "^1.3", optional = true}
markupsafe  = { version = "2.0.1", optional = true}
mkdocs-include-markdown-plugin  = { version = "^1.0.0", optional = true}
mkdocs-material  = { version = "^8.0", optional = true}
mkdocstrings = {version = "^0.19.0", extras = ["python"], optional = true }
mkdocs-material-extensions  = { version = "^1.0.1", optional = true}
twine  = { version = "^3.3.0", optional = true}
mkdocs-autorefs = {version = "^0.3", optional = true}
pre-commit = {version = "^2.12.0", optional = true}
toml = {version = "^0.10.2", optional = true}
bump2version = {version = "^1.0.1", optional = true}
jupyter = {version = ">=1.0", optional = true}
azure-functions  = ">1.0,<2.0"
azure-datalake-utils = ">=0.5.7,<1.0"
great-expectations = ">=0.17,<1.0"
python-dotenv = ">=1.0.0,<2.0"

[tool.poetry.extras]
test = [
    "pytest",
    "black",
    "isort",
    "flake8",
    "flake8-docstrings",
    "pytest-cov",
    "ipykernel",
    "markupsafe",
    "Jinja2"
    ]

dev = ["tox", "pre-commit", "virtualenv", "pip", "twine", "toml", "bump2version", "jupyter", "python-dotenv"]

doc = [
    "mkdocs",
    "mkdocs-include-markdown-plugin",
    "mkdocs-material",
    "mkdocstrings",
    "mkdocs-material-extension",
    "mkdocs-autorefs"
    ]

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py38', 'py39', 'py310']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
# you can skip files as below
#skip_glob = docs/conf.py

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
